feature(users sign up endpoint): refactored sisgn up endpoint add enviroment variable refactored response format
