+++ .babelrc
@@ -1,3 +1,6 @@
 {
-    "presets": ["@babel/preset-env"]
+    "presets": ["@babel/preset-env"],
+    "plugins": [
+        "@babel/transform-runtime",
+      ]
 }
