+++ server/index.js
@@ -1,7 +1,9 @@
 /* eslint-disable no-console */
 import express from 'express';
 import bodyParser from 'body-parser';
-import UserRoute from './routes/user.routes';
+import UserRoute from './routes/user.auth.routes';
+
+require('dotenv').config();
 
 const app = express();
 
@@ -13,6 +15,6 @@ app.use(bodyParser.json());
 // Routes
 app.use('/api/v1/auth', UserRoute);
 
-const PORT = process.env.PORT || 3000;
+const { PORT } = process.env;
 
 app.listen(PORT, () => console.log(`App is running from PORT ${PORT}`));
