+++ server/routes/user.auth.routes.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import UsersControllers from '../controllers/users.controllers';
+import UsersControllers from '../controllers/users.auth.controllers';
 import UserValidations from '../middlewares/validations/users.validations';
 
 const router = express.Router();
